+++ src/main/java/com/example/repositories/RencontreRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RencontreRepository extends CrudRepository<Rencontre, Long>{}
+public interface RencontreRepository extends CrudRepository<Rencontre, Integer>{}
